requirements.txt
Flask
opencv-python-headless
numpy
easyocr
imutils
Pillow
gunicorn
